Fix return value - return value is only set by `return` not by the result of the last expression. Lint’ing. Some restructuring to reduce the length of the asteval.py file.
